requirements.txt
mesop
vertexai
huggingface_hub
python-dotenv
diffusers
transformers
optimum.quanto
torch